+++ changelog.md
@@ -1,3 +1,7 @@
+### 2023-05-31 三 11:42:38  
+- 能加 my_snippets cpp.snippets
+- bits/stdc++.h macos 不存在解决方案 https://www.cnblogs.com/lipoicyclic/p/14623307.html
+
 ### 2023-05-30 二 11:10:43   
 - 增加插件
     - Comment.nvim 注释插件